move dataset logic to separate module
